+++ app/src/main/java/ua/com/radiokot/photoprism/features/viewer/logic/DefaultVideoPlayerFactory.kt
@@ -45,6 +45,7 @@ class DefaultVideoPlayerFactory(
                                 )
                         )
                 )
+                    .fixClippingSources()
             )
             .build()
 }